Added save function to custom hook for more central functionality
